+++ extensions/in-memory/assetindex-memory/src/test/java/org/eclipse/dataspaceconnector/metadata/memory/InMemoryDataAddressResolverTest.java
@@ -17,7 +17,7 @@ class InMemoryDataAddressResolverTest {
 
     @BeforeEach
     void setUp() {
-        resolver = new InMemoryAssetIndex(new AssetPredicateConverter());
+        resolver = new InMemoryAssetIndex();
     }
 
     @Test